requirements.txt
fastapi
uvicorn
jinja2
python-multipart
motor
pandas
pymongo
pydantic[dotenv]
python-dotenv
pydantic[email]
passlib[bcrypt]
pyjwt
passlib
secrets
scipy
scikit-learn
tqdm
sqlalchemy
requests
pymysql
fuzzywuzzy
python-Levenshtein
pycountry
